Fix native build, execute function body and support return via status codes
